Move LegacyReportingUtils to junit-platform-reporting
Move LegacyReportingUtils from junit-platform-launcher to
junit-platform-reporting as being more appropriate. Old implementation
in the launcher package left in place for compatibility with 3rd party code,
but deprecated for future removal.

Implements #1978.
